pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emery-cli"
version = "1.0.0"
description = "Modern file upload CLI with drag-and-drop support"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Emery Team", email = "cli@emery.local"}]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "typer==0.9.0",
    "rich==13.7.0",
    "GitPython==3.1.40",
    "click==8.1.7",
]

[project.scripts]
emery = "emery_cli.main:main"

[project.urls]
Homepage = "https://github.com/JuanJuanCtrl/Emery"
Repository = "https://github.com/JuanJuanCtrl/Emery.git"

[tool.setuptools]
packages = ["emery_cli"]
